fix(gateway): accept Anthropic thinking blocks
The native /v1/messages endpoint returned 400 for any request whose
history contained `thinking`/`redacted_thinking` content blocks (sent
by Claude Code and other extended-thinking clients), breaking
multi-turn sessions once the model actually reasoned. Accept these
blocks in the message schema and strip them during the
Anthropic->OpenAI transform so they never reach the inner completions
endpoint, which also rejects them. Reasoning for the next turn is
driven by the request's thinking/output_config controls, so dropping
replayed blocks is lossless.

Extract the message schema and transform into transform-messages.ts
(free of the app import) and add unit tests covering acceptance,
stripping, and the existing block types.

Fixes #2508.
