+++ apps/gateway/src/anthropic/anthropic.ts
@@ -13,80 +13,14 @@ import { logger, toError } from "@llmgateway/logger";
 
 import { buildAnthropicErrorEvent } from "./streaming-error-translation.js";
 import { mapAnthropicThinkingToReasoning } from "./thinking-to-reasoning.js";
+import {
+	anthropicMessageSchema,
+	anthropicMessagesToOpenai,
+} from "./transform-messages.js";
 
 import type { ServerTypes } from "@/vars.js";
 
 export const anthropic = new OpenAPIHono<ServerTypes>();
-
-const anthropicMessageSchema = z.object({
-	role: z.enum([
-		"system",
-		"developer",
-		"user",
-		"assistant",
-		"tool",
-		"function",
-	]),
-	content: z.union([
-		z.string(),
-		z.array(
-			z.union([
-				z.object({
-					type: z.literal("text"),
-					text: z.string(),
-					cache_control: z
-						.object({
-							type: z.enum(["ephemeral"]),
-							ttl: z.enum(["5m", "1h"]).optional(),
-						})
-						.optional(),
-				}),
-				z.object({
-					type: z.literal("image"),
-					source: z.object({
-						type: z.literal("base64"),
-						media_type: z.string(),
-						data: z.string(),
-					}),
-				}),
-				z.object({
-					type: z.literal("tool_use"),
-					id: z.string(),
-					name: z.string(),
-					input: z.record(z.unknown()),
-				}),
-				z.object({
-					type: z.literal("tool_result"),
-					tool_use_id: z.string(),
-					content: z.union([z.string(), z.array(z.unknown())]).optional(),
-					is_error: z.boolean().optional(),
-				}),
-			]),
-		),
-	]),
-	// OpenAI message properties
-	tool_call_id: z.string().optional(),
-	name: z.string().optional(),
-	tool_calls: z
-		.array(
-			z.object({
-				id: z.string(),
-				type: z.literal("function"),
-				function: z.object({
-					name: z.string(),
-					arguments: z.string(),
-				}),
-			}),
-		)
-		.optional(),
-	function_call: z
-		.object({
-			id: z.string().optional(),
-			name: z.string(),
-			arguments: z.union([z.string(), z.record(z.unknown())]),
-		})
-		.optional(),
-});
 
 const anthropicToolSchema = z.object({
 	name: z.string(),
@@ -319,208 +253,10 @@ anthropic.openapi(messages, async (c) => {
 		}
 	}
 
-	// Transform messages using the approach from claude-code-proxy
-	for (const message of anthropicRequest.messages) {
-		// Handle tool role → convert to OpenAI tool format
-		if (message.role === "tool") {
-			openaiMessages.push({
-				role: "tool",
-				content:
-					typeof message.content === "string"
-						? message.content
-						: JSON.stringify(message.content),
-				tool_call_id: message.tool_call_id,
-			});
-			continue;
-		}
-
-		// Handle function role → convert to OpenAI tool format (legacy)
-		if (message.role === "function") {
-			openaiMessages.push({
-				role: "tool",
-				content: message.content,
-				tool_call_id: message.tool_call_id ?? message.name,
-			});
-			continue;
-		}
-
-		// Handle assistant messages with tool_calls (OpenAI format)
-		if (message.role === "assistant" && message.tool_calls) {
-			openaiMessages.push({
-				role: message.role,
-				content: message.content || "",
-				tool_calls: message.tool_calls,
-			});
-			continue;
-		}
-
-		// Handle assistant messages with function_call (legacy OpenAI format)
-		if (message.role === "assistant" && message.function_call) {
-			const toolCalls = [
-				{
-					id:
-						message.function_call.id ??
-						`call_${Math.random().toString(36).substring(2, 10)}`,
-					type: "function" as const,
-					function: {
-						name: message.function_call.name,
-						arguments:
-							typeof message.function_call.arguments === "string"
-								? message.function_call.arguments
-								: JSON.stringify(message.function_call.arguments),
-					},
-				},
-			];
-
-			openaiMessages.push({
-				role: message.role,
-				content: message.content || "",
-				tool_calls: toolCalls,
-			});
-			continue;
-		}
-
-		// Handle assistant messages with tool_use blocks (native Anthropic format)
-		if (
-			message.role === "assistant" &&
-			Array.isArray(message.content) &&
-			message.content.some((block) => block.type === "tool_use")
-		) {
-			const toolCalls = message.content
-				.filter((block) => block.type === "tool_use")
-				.map((block) => ({
-					id: block.id,
-					type: "function" as const,
-					function: {
-						name: block.name,
-						arguments: JSON.stringify(block.input),
-					},
-				}));
-
-			const textContent = message.content
-				.filter((block) => block.type === "text")
-				.map((block) => block.text)
-				.join("");
-
-			openaiMessages.push({
-				role: message.role,
-				content: textContent || "",
-				tool_calls: toolCalls,
-			});
-			continue;
-		}
-
-		// Handle user messages with tool_result blocks (native Anthropic format)
-		if (
-			message.role === "user" &&
-			Array.isArray(message.content) &&
-			message.content.some((block) => block.type === "tool_result")
-		) {
-			// Group tool_result blocks by tool_use_id to avoid duplicates
-			const toolResults = new Map<string, any[]>();
-			for (const block of message.content) {
-				if (block.type === "tool_result") {
-					const toolUseId = block.tool_use_id;
-					if (!toolResults.has(toolUseId)) {
-						toolResults.set(toolUseId, []);
-					}
-					toolResults.get(toolUseId)!.push(block);
-				}
-			}
-
-			// Convert each unique tool_use_id to a single tool message
-			for (const [toolUseId, blocks] of toolResults) {
-				// Combine content from all blocks with the same tool_use_id
-				const combinedContent = blocks
-					.map((block) =>
-						typeof block.content === "string"
-							? block.content
-							: JSON.stringify(block.content),
-					)
-					.join("\n");
-
-				openaiMessages.push({
-					role: "tool",
-					content: combinedContent,
-					tool_call_id: toolUseId,
-				});
-			}
-
-			// Handle any remaining text content as a user message
-			const textContent = message.content
-				.filter((block) => block.type === "text")
-				.map((block) => block.text)
-				.join("");
-
-			if (textContent) {
-				openaiMessages.push({
-					role: "user",
-					content: textContent,
-				});
-			}
-			continue;
-		}
-
-		// Handle regular messages and multi-modal content
-		if (Array.isArray(message.content)) {
-			// Check if this is complex multi-modal content that should be flattened
-			const hasOnlyText = message.content.every(
-				(block) => block.type === "text",
-			);
-			const hasAnyCacheControl = message.content.some(
-				(block) => block.type === "text" && block.cache_control,
-			);
-
-			if (hasOnlyText && !hasAnyCacheControl) {
-				// For text-only content with no cache markers, flatten to a simple
-				// string to avoid content type issues.
-				const textContent = message.content
-					.filter((block) => block.type === "text")
-					.map((block) => block.text)
-					.join("");
-
-				openaiMessages.push({
-					role: message.role,
-					content: textContent,
-				});
-			} else {
-				// For multi-modal content, or text content with cache_control markers,
-				// transform blocks while preserving cache_control so the inner
-				// completions path can forward it to Anthropic.
-				const content = message.content.map((block) => {
-					if (block.type === "text" && block.text) {
-						return {
-							type: "text",
-							text: block.text,
-							...(block.cache_control && {
-								cache_control: block.cache_control,
-							}),
-						};
-					}
-					if (block.type === "image" && block.source) {
-						return {
-							type: "image_url",
-							image_url: {
-								url: `data:${block.source.media_type};base64,${block.source.data}`,
-							},
-						};
-					}
-					return block;
-				});
-
-				openaiMessages.push({
-					role: message.role,
-					content,
-				});
-			}
-		} else {
-			// Simple string content
-			openaiMessages.push({
-				role: message.role,
-				content: message.content,
-			});
-		}
-	}
+	// Transform conversation messages (Anthropic Messages format -> OpenAI),
+	// stripping extended-thinking blocks the inner completions endpoint can't
+	// accept. See ./transform-messages.ts.
+	openaiMessages.push(...anthropicMessagesToOpenai(anthropicRequest.messages));
 
 	// Transform tools if provided
 	let openaiTools;
